pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arpi-server"
dynamic = ["version"]
description = "The backend service of the ArPI security system running on the Raspberry PI"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ArPI Home Security"}
]
keywords = ["security", "home-automation", "raspberry-pi", "alarm-system"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Home Automation",
    "Topic :: Security",
]

dependencies = [
    "bcrypt~=5.0",
    "flask~=3.1",
    "flask-cors~=6.0",
    "flask-migrate~=4.1",
    "flask-sqlalchemy~=3.1",
    "fastmcp~=3.0",
    "gunicorn~=25.1",
    "noipy~=1.5.3",
    "paho-mqtt~=2.1",
    "psycopg2-binary~=2.9",
    "python-crontab~=2.7.0",
    "python-gsmmodem-new~=0.13.0",
    "python-jose~=3.5",
    "python-socketio~=5.14",
    "stringcase~=1.2.0",
    "uvicorn~=0.42",
]

[project.optional-dependencies]
dev = [
    "black~=23.12",
    "cryptography~=46.0",
    "rope~=1.13",
]
simulator = [
    "textual~=5.3",
    "textual-dev",
]
device = [
    "pywiegand==0.5.2",
]

[project.urls]
Homepage = "https://www.arpi-security.info/"
Documentation = "https://docs.arpi-security.info/"
Demo = "https://demo.arpi-security.info/"
Repository = "https://github.com/ArPIHomeSecurity/arpi_server"

[project.scripts]
argus-certbot = "bin.certbot:cli_main"
argus-clock = "bin.clock:cli_main"
argus-data = "bin.data:cli_main"
argus-dyndns = "bin.dyndns:cli_main"
argus-hash = "bin.hash:cli_main"
argus-secret = "bin.secret:cli_main"
argus-simulator = "bin.simulator:main"
argus-ssh-keymanager = "bin.ssh_keymanager:cli_main"
argus-ssh-service = "bin.ssh_service:cli_main"
argus-tester = "bin.tester:cli_main"
argus-user = "bin.user:cli_main"
argus-upgrade = "bin.upgrade:cli_main"
argus-status = "installer.cli:status_standalone"

[tool.setuptools]
packages = {find = {where = ["src"]}}

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
server = ["version.json"]
mcp_server = ["prompts/*.txt"]

[tool.setuptools.data-files]
"etc/arpi-server/" = ["config.env"]
"etc/arpi-server" = ["src/server/gunicorn.conf.py"]
"etc/arpi-monitor" = ["src/monitor/gunicorn.conf.py"]
"share/arpi-server/migrations" = [
    "migrations/alembic.ini",
    "migrations/env.py",
    "migrations/README",
    "migrations/script.py.mako",
]
"share/arpi-server/migrations/versions" = ["migrations/versions/*.py"]

[tool.setuptools.dynamic]
version = {attr = "server.version.__package_version__"}

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
